main.pyを切り出し、srcディレクトリに各ファイルを格納した
