+++ backend/Dockerfile
@@ -6,13 +6,14 @@ COPY --from=ghcr.io/astral-sh/uv:0.7.16 /uv /uvx /bin/
 
 COPY . /app
 # 作業ディレクトリを設定
-WORKDIR /app
+WORKDIR /app/src
 
 # 依存関係ファイルをコピーしてインストール
 RUN uv sync
 
 # 仮想環境を設定
 ENV PATH="/app/.venv/bin:$PATH"
+ENV PYTHONPATH="/app:$PYTHONPATH"
 
 # FastAPIサーバーを起動
-CMD ["/bin/bash", "-c", "source .venv/bin/activate && uvicorn main:app --host 0.0.0.0 --port 8000"]
+CMD ["/bin/bash", "-c", "source /app/.venv/bin/activate && uvicorn src.main:app --host 0.0.0.0 --port 8000"]
