+++ backend/src/models.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class Item(BaseModel):
+    url: str
